Support custom serializers
Alamofire supports custom response serializers:

* https://littlebitesofcocoa.com/94-custom-alamofire-response-serializers
* https://github.com/Alamofire/Alamofire/blob/c8ffe77f0b7bcdbacdbbc6a13d5de18e53c08ae4/Documentation/AdvancedUsage.md#creating-a-custom-response-serializer

I have a use case where I've written one of these, and would like to use it with PromiseKit. When I add an extension to support my custom serializer, I'm unable to initialize the PMKAlamofireDataResponse because of the protection level. This is the error:

```
'PMKAlamofireDataResponse' initializer is inaccessible due to 'fileprivate' protection level
```